Add account deletion confirmation request
